Cambiamos a Fetch Eager en los Beans Ejemplar, Libro y Usuario
- Añadimos también la anotación FetchMode Select, para permitir cargas anidadas
- Añadido listado de préstamos a la vista usuario/ver y ajustes en libro/ver y prestamo/index
- Mejorado formulario de nuevo Préstamo, para aceptar Usuario y/o Libro preseleccionado
- Eliminamos la ruta prestamo/ver, ya que finalmente no será necesaria
- Finalizado TODO: Cálculos en el Bean de Ejemplar
- Finalizado TODO: Cálculos en el Bean de Libro

TODO
- Validaciones del Bean de Usuario
- Establecer mensajes flash (Préstamos y Usuarios)

+++ src/main/java/org/alexandrialibrary/spring/controller/UsuarioController.java
@@ -111,6 +111,14 @@ public class UsuarioController extends AbstractController {
 	public String eliminar(@PathVariable Long id, Model model) {
 		logger.info("Iniciamos usuario/eliminar/{id} [GET]");
 
+		Usuario usuario = usuarioDAO.getUsuario(id);
+		if (usuario.hasPrestamosPendientes()) {
+			logger.info("Tiene préstamos pendientes, no se puede eliminar.");
+			logger.info("Redireccionamos a usuario/ver/{id} [GET]");
+			return String.format("redirect:/usuario/ver/%d", id);
+		}
+
+		logger.info("No tiene préstamos pendientes, eliminamos el usuario.");
 		usuarioDAO.delete(id);
 
 		logger.info("Redireccionamos a usuario/listado [GET]");
